chore: move wizards folder under screens folder
