+++ ts/navigation/AuthenticationNavigator.tsx
@@ -26,10 +26,10 @@ import CieWrongCardScreen from "../screens/authentication/cie/CieWrongCardScreen
 import ActivateNfcScreen from "../screens/authentication/cie/ActivateNfcScreen";
 import AuthErrorScreen from "../screens/authentication/AuthErrorScreen";
 import OptInScreen from "../screens/authentication/OptInScreen";
-import CieIdWizard from "../features/cie/wizards/CieIdWizard";
-import CiePinWizard from "../features/cie/wizards/CiePinWizard";
-import SpidWizard from "../features/cie/wizards/SpidWizard";
-import IDActivationWizard from "../features/cie/wizards/IDActivationWizard";
+import CieIdWizard from "../features/cie/screens/wizards/CieIdWizard";
+import CiePinWizard from "../features/cie/screens/wizards/CiePinWizard";
+import SpidWizard from "../features/cie/screens/wizards/SpidWizard";
+import IDActivationWizard from "../features/cie/screens/wizards/IDActivationWizard";
 import { AuthenticationParamsList } from "./params/AuthenticationParamsList";
 import ROUTES from "./routes";
 import CloseButton from "./components/CloseButton";
